+++ packages/process_run/lib/shell.dart
@@ -23,6 +23,12 @@ export 'package:process_run/dartbin.dart'
 export 'package:process_run/src/api/shell_common.dart' show ShellOptions;
 // We reuse io sharedStdIn definition.
 export 'package:process_run/src/io/shared_stdin.dart' show sharedStdIn;
+export 'package:process_run/src/io/user_shell_alias_io.dart'
+    show
+        userShellAliasEnabled,
+        userShellAliasShellPath,
+        userShellAliasShellPathOverride,
+        userShellAliasSupportedShells;
 export 'package:process_run/src/shell_utils.dart'
     show
         userHomePath,